Changed flight gps data to float

+++ TargetsMicroservice/Requests/FlightBeginRequest.cs
@@ -6,9 +6,9 @@
         public int TeamID { get; set; }
         public long FlightID { get; set; }
         public int PlatoonID { get; set; }
-        public int X { get; set; }
-        public int Y { get; set; }
-        public int Z { get; set; }
+        public float X { get; set; }
+        public float Y { get; set; }
+        public float Z { get; set; }
         public DateTime BeginTime { get; set; }
         public string? Comment { get; set; }
     }
